trying with size_t size

+++ 0x1E-search_algorithms/search_algos.h
@@ -5,6 +5,6 @@
 
 int linear_search(int *array, size_t size, int value);
 int binary_search(int *array, size_t size, int value);
-void print(int *array, int left, int right);
+void print(int *array, size_t left, size_t right);
 
 #endif
